GreedykDPPSampling/requirements.txt
dppy==0.3.2
emukit==0.4.10
GPy==1.12.0
matplotlib==3.8.0
numpy==1.22.1
pandas==1.1.5
scikit_learn==1.3.1
scikits.talkbox==0.2.5
scipy==1.11.3
tueplots==0.0.7
